Move function converting to a separate file
